fix 5 bugs: helper context skip logic, malformed cost_t, score() dtype, GAE O(n²), and fallback question count
1. _extract_helper_context: fixed skip state machine — now correctly
   skips the main function body and includes everything else
2. Malformed output cost_t: changed from 0.0 to 1.0 — malformed turns
   are no longer "free"
3. score() hidden state: added .float() to match generate() behavior
4. GAE: replaced insert(0,...) with append+reverse for O(n) instead of O(n²)
5. Fallback question count: default to 1 instead of counting ? in the
   simulator's response

+++ src/models/agent.py
@@ -262,8 +262,8 @@ class Agent:
             attention_mask=attention_mask,
             output_hidden_states=True,
         )
-        # Hidden state at last prompt token
-        state_hidden = policy_out.hidden_states[-1][0, prompt_len - 1, :]  # (hidden_dim,)
+        # Hidden state at last prompt token (float32 to match value heads)
+        state_hidden = policy_out.hidden_states[-1][0, prompt_len - 1, :].float()  # (hidden_dim,)
 
         # Log-probs of action tokens
         logits = policy_out.logits[0, prompt_len - 1: -1, :]  # (action_len, vocab)
